rework chat ui (wip)
