+++ client/package.json
@@ -39,7 +39,7 @@
 	"type": "module",
 	"dependencies": {
 		"@popperjs/core": "^2.11.8",
-		"bits-ui": "^0.21.9",
+		"bits-ui": "^0.21.10",
 		"classnames": "^2.3.2",
 		"clsx": "^2.1.1",
 		"flowbite": "^2.2.1",
