+++ client/src/lib/components/chat/message/components/table.svelte
