+++ client/src/lib/components/ui/sheet/sheet-content.svelte
@@ -37,11 +37,5 @@
 		{...$$restProps}
 	>
 		<slot />
-		<SheetPrimitive.Close
-			class="absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-secondary"
-		>
-			<X class="h-4 w-4" />
-			<span class="sr-only">Close</span>
-		</SheetPrimitive.Close>
 	</SheetPrimitive.Content>
 </SheetPortal>
